Rename package to teamR.
