+++ src/main/java/edu/wpi/teamR/App.java
@@ -1,7 +1,7 @@
-package edu.wpi.teamname;
+package edu.wpi.teamR;
 
-import edu.wpi.teamname.navigation.Navigation;
-import edu.wpi.teamname.navigation.Screen;
+import edu.wpi.teamR.navigation.Navigation;
+import edu.wpi.teamR.navigation.Screen;
 import java.io.IOException;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
