+++ src/main/java/edu/wpi/teamR/controllers/ExampleController.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamname.controllers;
+package edu.wpi.teamR.controllers;
 
 import java.io.IOException;
 import java.sql.*;
