+++ src/main/java/edu/wpi/teamR/controllers/RootController.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamname.controllers;
+package edu.wpi.teamR.controllers;
 
 import javafx.fxml.FXML;
