+++ src/main/java/edu/wpi/teamR/controllers/ServiceRequestController.java
@@ -1,7 +1,7 @@
-package edu.wpi.teamname.controllers;
+package edu.wpi.teamR.controllers;
 
-import edu.wpi.teamname.navigation.Navigation;
-import edu.wpi.teamname.navigation.Screen;
+import edu.wpi.teamR.navigation.Navigation;
+import edu.wpi.teamR.navigation.Screen;
 import io.github.palexdev.materialfx.controls.MFXButton;
 import javafx.fxml.FXML;
