+++ src/main/java/edu/wpi/teamR/navigation/Screen.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamname.navigation;
+package edu.wpi.teamR.navigation;
 
 public enum Screen {
   ROOT("views/Root.fxml"),
