+++ src/main/resources/edu/wpi/teamR/images/Brigham_and_Womens_Hospital_logo.svg.png
