+++ src/main/resources/edu/wpi/teamR/images/TriangleArrow-Right.svg.png
