+++ src/main/resources/edu/wpi/teamR/images/long_triangle.png
