+++ src/main/resources/edu/wpi/teamR/simplelogger.properties
