+++ src/test/java/edu/wpi/teamR/DefaultTest.java
@@ -2,7 +2,7 @@
 /* DO NOT DELETE THIS TEST */
 /*-------------------------*/
 
-package edu.wpi.teamname;
+package edu.wpi.teamR;
 
 import org.junit.jupiter.api.Test;
